Moving GetUserInfo to the interface
